Crate spectator

spectator

Spectator is a dedicated GUI program which displays the progress of an optimization run. The data is fed to spectator by the argmin-observer-spectator.

Installation

cargo install spectator

Library

This crate can also be used as a library and exposes the Message type used to encode data sent to specator and DEFAULT_PORT which defines the default port used by spectator.

License

Licensed under either of

Apache License, Version 2.0, (LICENSE-APACHE or http://www.apache.org/licenses/LICENSE-2.0)

MIT License (LICENSE-MIT or http://opensource.org/licenses/MIT)

at your option.

Contribution

Unless you explicitly state otherwise, any contribution intentionally submitted for inclusion in the work by you, as defined in the Apache-2.0 license, shall be dual licensed as above, without any additional terms or conditions.

Enums

Message

Enum used to encode information sent to spectator.

Constants

DEFAULT_PORT

Default port used by spectator